+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [1.3.0] - 2026-02-19
+Fixes interoperability with Zotero v8 + BetterBibtex, but not backward
+compatible.
+### Changed
+- Update queries for the new Zotero SQLite DB format for citation keys.
+- Move Better Bibtex mapping to use the new Zotero SQLite citation key table.
+
+
 ## [1.2.0] - 2026-01-04
 ### Changed
 - Switch build tools to [pixi].
@@ -245,7 +253,8 @@ Downstream moderate risk update release.
 - Initial version
 
 
-[Unreleased]: https://github.com/plandes/zotsite/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/plandes/zotsite/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/plandes/zotsite/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/plandes/zotsite/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/plandes/zotsite/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/plandes/zotsite/compare/v0.9.1...v1.0.0